Fix lexer for "----" (<hr/>) case
The point is that there may be another character in the string.
For example, a space or a line break.

Because we check the whole string, not the first 4 characters - we can ignore the variant where it can be "----\s" or "----\r\n" and make the wrong markup.

Alternatively: use `startsWith("----")`, but this is a bit more expensive operation.