Don't exit when removing socket fails
Prior to migrating to slog, errors when removing the socket were
ignored.
This is fine and desirable: the most likely error condition is that the
socket did not exist in the first place. Exiting on error in that case
effectively prevents Mycorrhiza to be used with unix sockets.
If the removal fails for another reason, then starting the server will
fail, so logging a warning is sufficient for troubleshooting.